move all blog content into one folder
