+++ fresh.config.ts
@@ -7,7 +7,7 @@ import securityHeaders from "./plugins/security_headers.ts";
 import welcomePlugin from "./plugins/welcome.ts";
 import type { FreshConfig } from "$fresh/server.ts";
 import { ga4Plugin } from "https://deno.land/x/fresh_ga4@0.0.4/mod.ts";
-import { blog } from "./plugins/blog.ts";
+import { blog } from "./plugins/blog/blog.ts";
 
 export default {
   plugins: [
