+++ plugins/blog/routes/blog/[slug].tsx
@@ -1,7 +1,7 @@
 // Copyright 2023-2024 the Deno authors. All rights reserved. MIT license.
 import { defineRoute } from "$fresh/server.ts";
 import { CSS, render } from "$gfm";
-import { getPost } from "@/utils/posts.ts";
+import { getPost } from "../../utils/posts.ts";
 import Head from "@/components/Head.tsx";
 import Share from "@/components/Share.tsx";
